Added async function/method buyNFT(nft)
This uses Web3Modal to connect to the users wallet. At the bottom of the function/method loadNFTs() should show us the NFT's that have not yet been purchased. Below that function: if there are no nfts available, a message is shown which tells the user that there are no NFT's available in the Marketplace yet.